Change type of num_sequences field to integer

+++ webfront/models/interpro_new.py
@@ -65,7 +65,7 @@ class EntryAnnotation(models.Model):
     type = models.CharField(max_length=32)
     value = models.BinaryField()
     mime_type = models.CharField(max_length=32)
-    num_sequences = models.FloatField(null=True)
+    num_sequences = models.IntegerField(null=True)
 
 
 class Protein(models.Model):
